Move sandbox internals into runtimehost service
New interface:

- `services/runtimehost.RunSandboxSubcommand(args)` is the narrow daemon-facing
  entrypoint for the hidden sandbox helper command.
- The sandbox implementation lives behind the service-local internal package at
  `services/runtimehost/internal/sandbox`.

Usage examples:

```go
if err := runtimehost.RunSandboxSubcommand(args); err != nil {
    return err
}
```

```go
policy := &sandbox.Policy{AllowedHosts: egressPolicy.AllowedHosts}
wrapped, cleanup, err := sandbox.Wrap(policy, cmd)
```

The second example is intentionally only available inside `services/runtimehost`,
keeping process isolation details inside the runtime host service boundary while
allowing the daemon to dispatch the helper subcommand through the public wrapper.
