+++ gestaltd/services/runtimehost/sandbox.go
@@ -1,0 +1,8 @@
+package runtimehost
+
+import "github.com/valon-technologies/gestalt/server/services/runtimehost/internal/sandbox"
+
+// RunSandboxSubcommand runs the runtimehost sandbox helper command.
+func RunSandboxSubcommand(args []string) error {
+	return sandbox.RunSubcommand(args)
+}
